Release 1.6.0

+++ HayRackController/src/main/java/gift/goblin/HayRackController/controller/TimeTableController.java
@@ -5,18 +5,20 @@
  */
 package gift.goblin.HayRackController.controller;
 
-import gift.goblin.HayRackController.database.event.model.ScheduledShutterMovement;
+import gift.goblin.HayRackController.database.model.event.ScheduledShutterMovement;
 import gift.goblin.HayRackController.service.io.WebcamDeviceService;
 import gift.goblin.HayRackController.service.scheduled.SchedulerJobService;
-import gift.goblin.HayRackController.database.event.ScheduledShutterMovementService;
+import gift.goblin.HayRackController.service.event.ScheduledShutterMovementService;
+import gift.goblin.HayRackController.database.embedded.repo.event.ScheduledShutterMovementRepository;
 import gift.goblin.HayRackController.service.tools.DateAndTimeUtil;
-import gift.goblin.HayRackController.view.model.ScheduledShutterMovementDto;
+import gift.goblin.HayRackController.controller.model.ScheduledShutterMovementDto;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import org.quartz.JobDetail;
 import org.quartz.Scheduler;
@@ -32,8 +34,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 
 /**
  * Controller for the timetable view, where you can see the scheduled timings
@@ -48,6 +48,9 @@ public class TimeTableController {
 
     @Autowired
     private ScheduledShutterMovementService scheduledShutterMovementService;
+    
+    @Autowired
+    ScheduledShutterMovementRepository repo;
 
     @Autowired
     private SchedulerJobService schedulerJobService;
@@ -66,12 +69,8 @@ public class TimeTableController {
 
         List<ScheduledShutterMovement> scheduledMovements = scheduledShutterMovementService.readAllStoredShutterMovementSchedules();
 
-        for (ScheduledShutterMovement actSchedule : scheduledMovements) {
-            logger.info(actSchedule.toString());
-        }
-
         List<ScheduledShutterMovementDto> shutterMovementDtos = scheduledMovements.stream().map((ScheduledShutterMovement s) -> new ScheduledShutterMovementDto(s.getId().toString(),
-                s.isIsActive(), s.getFeedingStartTime().toString(), s.getFeedingDuration().toString(), s.getCreatedBy(), s.getCreatedAt().toString()))
+                s.getFeedingStartTime().toString(), s.getFeedingDuration().toString(), s.getCreatedBy(), s.getCreatedAt().toString()))
                 .collect(Collectors.toList());
 
         model.addAttribute("scheduledMovements", shutterMovementDtos);
@@ -91,26 +90,38 @@ public class TimeTableController {
 
         // Register scheduled job
         registerStartFeedingJob(newMovement.getFeedingStartTime(), newShutterMovementId);
+        
+        model.addAttribute("added_time", newMovement.getFeedingStartTime());
+        model.addAttribute("success_message", "timetable.addNew.success");
 
         return renderTimetable(model);
     }
 
     @GetMapping(value = "/timetable/delete/{id}")
     public String deleteEntry(@PathVariable("id") String id, Model model) {
-
-        scheduledShutterMovementService.deleteScheduledMovement(Long.valueOf(id));
-        schedulerJobService.deleteStartFeedingJob(Integer.valueOf(id));
+        
+        try {
+            Optional<ScheduledShutterMovement> optScheduledShutterMovement = repo.findById(Long.valueOf(id));
+            if (optScheduledShutterMovement.isPresent()) {
+                LocalTime feedingStartTime = optScheduledShutterMovement.get().getFeedingStartTime();
+                
+                scheduledShutterMovementService.deleteScheduledMovement(Long.valueOf(id));
+                schedulerJobService.deleteStartFeedingJob(Integer.valueOf(id));
+                
+                model.addAttribute("deleted_time", feedingStartTime);
+                model.addAttribute("success_message", "timetable.delete.success");
+            }
+            
+        } catch (Exception e) {
+            logger.error("Exception thrown while try to delete scheduled entry with id: {}", id);
+        }
         
         return renderTimetable(model);
     }
 
     private void registerStartFeedingJob(LocalTime localTime, Long schedulerId) {
-
-        LocalDateTime nextExecutionDateTime = dateAndTimeUtil.getNextExecutionDateTime(localTime);
-
-        ZonedDateTime zdt = nextExecutionDateTime.atZone(ZoneId.systemDefault());
-        Date nextExecutionDate = Date.from(zdt.toInstant());
-
+        
+        Date nextExecutionDate = dateAndTimeUtil.getNextExecutionDate(localTime);
         JobDetail jobDetail = schedulerJobService.createStartFeedingJob(schedulerId.intValue());
         SimpleTrigger newTrigger = schedulerJobService.createStartFeedingTrigger(schedulerId.intValue(), nextExecutionDate, jobDetail);
